requirements.txt
pytest>=3.6.4
pytorch-transformers>=1.2.0
python==3.6.8
dask[dataframe]==1.2.2
azureml-train-automl==1.0.57
pydocumentdb>=2.3.3
numba>=0.38.1
py-xgboost<=0.80
cudatoolkit==9.2
scikit-learn>=0.19.0,<=0.20.3
scipy>=1.0.0
tensorflow==1.12.0
tqdm==4.31.1
pre-commit>=1.14.4
sklearn-crfsuite>=0.3.6
matplotlib>=2.2.2
pytorch==1.0.0
nltk>=3.4
black>=18.6b4
allennlp>=0.8.4
seqeval>=0.0.12
pyemd==0.5.1
spacy>=2.1.4
h5py>=2.8.0
cuda90
pandas>=0.24.2
cached-property==1.5.1
jupyter>=1.0.0
tensorflow-gpu==1.12.0
azureml-dataprep==1.1.8
pytorch-pretrained-bert>=0.6
gensim>=3.7.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
ipywebrtc==0.4.3
pip>=19.1.1
azureml-mlflow>=1.0.43.1
nteract-scrapbook>=0.2.1
azureml-sdk[automl,notebooks,contrib]==1.0.57
papermill>=1.0.1
pytorch>=1.0.0
pytorch-cpu>=1.0.0
numpy>=1.13.3
azureml-widgets==1.0.57
seaborn>=0.9.0
tensorflow-hub==0.5.0
ipykernel>=4.6.1